An alternative main loop where draw-screen is called in the backend on a page flip event
This was an experiment to see if it affected animation smoothness or not.

Didn't seem to have any effect.

The original render loop only sets a flag on the page flip event which is
checked for being false in main-loop-drm. With this setup we don't actually
know when we'll attempt to draw another frame...it will depend on receiving
some event on the event loop that requires a screen redraw.

By drawing the screen in the page flip handler I was wondering if it would
improve animation smoothness by immediatelly drawing the next screen. Perhaps
if the drawing is delayed we miss the next vblank and that was causing animation
issues. Again this commit seems to have had no effect.