Finally support resizing from borders
Currently this only supports resizing from bottom, right
and bottom right edges. This is because we also need to
update the surface x and y if resizing from other edges.

I did implement that but my naive implementation didn't
really play nice. The size of weston-terminal does not
increase in pixel increments, rather it needs a minimum
resize size in order to actually update its width and
height. If I just modify x with delta-x and y with delta-y
the bottom and right edges are not fixed as they should be
but rather "vibrate". I need to think about how to solve
this properly.